Move partitionReport to general utils.
